+++ examples/stm32l4/src/main.rs
@@ -79,13 +79,13 @@ fn main() -> ! {
     let mut lps25hb = LPS25HB::new(i2c_interface);
 
     // turn the sensor on 
-    lps25hb.sensor_power(Control::On).unwrap();
+    lps25hb.sensor_on(true).unwrap();
     
     // enable Block Data Update
-    lps25hb.bdu_config(Control::On).unwrap();
+    lps25hb.bdu_enable(true).unwrap();
 
     // set data rate to 7Hz
-    lps25hb.set_datarate(ODR::_7Hz).unwrap();
+    lps25hb.set_datarate(ODR::_1Hz).unwrap();
 
     loop {
             
@@ -93,11 +93,14 @@ fn main() -> ! {
         let temp = lps25hb.read_temperature().unwrap();            
         let press = lps25hb.read_pressure().unwrap();
 
+        let id = lps25hb.get_device_id().unwrap();
+
         // LED on
         led.set_high().ok();    
         
         // print data to serial
-        writeln!(tx, "temperature: {:.2}, pressure: {:.2}\r", temp, press).unwrap();
+        writeln!(tx, "temperature: {:.1} C, pressure: {:.1} hPa\r", temp, press).unwrap();
+        //writeln!(tx, "my name is: {}\r", id).unwrap();
         
         // wait a little
         delay.delay_ms(100 as u32);